migrate: TypeScript 마이그레이션, folder structure 수정
