+++ src/components/interface/Form.module.scss
@@ -1,5 +1,6 @@
 @import "@/styles/utils.scss";
 
+// Search
 .search_wrapper {
     position: relative;
 
@@ -40,6 +41,7 @@
     }
 }
 
+// Input
 .input[type="text"],
 .input[type="password"] {
     display: block;
@@ -76,53 +78,10 @@
     accent-color: $color-primary;
 }
 
+// Label
 .label {
 }
 
-.btn_default {
-    display: block;
-
-    border-radius: 10px;
-    font-size: 1rem;
-    font-weight: bold;
-}
-
-.btn_primary_stroke {
-    border: 2px solid $color-primary;
-    color: $color-primary;
-    background-color: #fff;
-}
-.btn_primary_fill {
-    color: #fff;
-    background-color: $color-primary;
-}
-
-.social_login_btn {
-    box-sizing: border-box;
-    flex-grow: 1;
-
-    display: flex;
-    align-items: center;
-    height: 55px;
-    width: 100%;
-
-    padding: 10px 20px;
-    border: 1px solid #DBDBDB;
-    border-radius: 10px;
-
-    background-color: #fff;
-
-    img{
-        height: 100%;
-        aspect-ratio: 1/1;
-        object-fit: contain;
-    }
-    p{
-        width: 100%;
-        color: #767676;
-        font-weight: bold;
-    }
-    &:hover{
-        cursor: pointer;
-    }
+// InputContainer
+.input_container {
 }
