+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 
-import App from "./App.jsx";
+import App from "./App.tsx";
 import "./styles/globals.scss";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
