+++ vite.config.ts
@@ -2,6 +2,8 @@ import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
 import path from "path";
 
+const __dirname = path.resolve();
+
 // https://vitejs.dev/config/
 export default defineConfig({
     plugins: [react()],
